refactor: Atualizar animações #25
