+++ ansi_utils.c
@@ -1,7 +1,5 @@
 #include <stdio.h>
 #include <string.h>
-#include <sys/ioctl.h>
-#include <unistd.h>
 
 #include "terminal_control.h"  // ✅ Inclui a declaração de enum Color e os protótipos
 
@@ -47,30 +45,6 @@ void ansi_print(int row, int col, const char *text) {
     move_cursor(row, col);
     printf("%s", text);
     fflush(stdout);
-}
-
-// Função para verificar mudanças no tamanho do terminal
-void update_terminal_size(int *last_rows, int *last_cols) {
-    struct winsize w;
-    if (ioctl(STDOUT_FILENO, TIOCGWINSZ, &w) == -1) {
-        return; // Falha ao pegar tamanho
-    }
-
-    // Se o tamanho do terminal mudou, atualiza as variáveis
-    if (*last_rows != w.ws_row || *last_cols != w.ws_col) {
-        *last_rows = w.ws_row;
-        *last_cols = w.ws_col;
-        // Redesenha a tela
-        clear_screen();
-    }
-}
-
-// Função para definir o tamanho do terminal
-void set_terminal_size(int rows, int cols) {
-    struct winsize size;
-    size.ws_row = rows;   // Número de linhas
-    size.ws_col = cols;   // Número de colunas
-    ioctl(STDOUT_FILENO, TIOCSWINSZ, &size);
 }
 
 // Ativa o negrito no terminal
