+++ include/terminal_control.h
@@ -26,7 +26,6 @@ enum Color {
 void move_cursor(int, int);
 void ansi_print(int, int, const char*);
 void clear_screen(void);
-void update_terminal_size(int*, int*);
 void hide_cursor(void);
 void show_cursor(void);
 void enable_bold(void);
@@ -39,5 +38,7 @@ char get_keypress(void);
 void disable_raw_mode(void);
 void enable_raw_mode(void);
 void restore_terminal(void);
+void set_terminal_size(int, int);
+void update_terminal_size(int*, int*);
 
 #endif
